gsa-contract-search-bot/govcon-scout/requirements.txt
flask
requests
python-dotenv
ollama
